Add support for logging through syslog
When using logs from syslog - and Hodel 3k and their ilk probably fall
into that, there's a bunch more data that is availabe, but isn't used.
This commit adds a module that decorates all of the other file formats
with support for syslog.

However, at the moment, the syslog module does not work nicely with
unioned regexps.