Add support for tracking with a PID
Getting results out of order is annoying.  This commit resolves that, by
tracking the current request on a per PID basis, if the line_definition
provides a PID, MUCH fewer requests get missed.

It's possible for memory usage to increaes greately if we get a lot of
unifinished (crashes) requests.